Add latest newhope Tor reference sources from the university repo.
These were taken from a private repo which includes a reference implementation
meant for Tor, i.e. where the generation of NewHope's `a` value is produced in
constant-time via oversampling from the PRNG and running the values through a
Batcher merge-sorting network.

This copy of the code was taken from the master branch, commit 7ede68d4.