build: Remove [root] section from top-level Cargo.lock file.
Newer versions of cargo recently removed the ability to understand the
`[root]` section of Cargo.lock files, which is particularly relevant when
either `cargo --frozen` and/or `cargo --locked` are specified, and the
failure/inability to download a new index will cause the whole build to fail
with "error: the lock file needs to be updated but --frozen was passed to
prevent this" (cf. https://github.com/rust-lang/cargo/issues/4563).

 * REMOVE `[root]` section from our top-level `Cargo.lock` file.
 * FIXES #24608: https://bugs.torproject.org/24608