rust: Fix module publicity/exportation in crypto_rand_f64.
This fixes the issues with unused code warnings (these were due to `pub use`ing
modules and then only using the things inside the modules).  It also fixes the
weirdness with the compiler not being able to find `F64_INTEGER_MAX_AS_U64`
unless it was explicitly imported from the top level.